requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Machine Learning & Computer Vision
tensorflow==2.15.0
opencv-python==4.8.1.78
numpy==1.24.3
pillow==10.1.0
scikit-learn==1.3.2
scikit-image==0.22.0

# OCR
pytesseract==0.3.10

# Image Processing
imutils==0.5.4
pdf2image==1.16.3

# HTTP Client
httpx==0.25.1
aiohttp==3.9.1

# Utilities
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0
